Restructure auth components
