+++ src/components/SignUpType/style.module.css
