+++ src/screens/Auth/SignUpType.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SignUpType from '../../components/Auth/SignUpType';
+import SignUpType from '../../components/SignUpType';
 
 export default () => {
   return <SignUpType />;
